+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - You can now choose the formatting characters that should be used by the formatting commands.
 - You can now change the pattern used for generating new file names, and omit being asked for filenames in general.
 - Zettlr now tries to escape the input you provide for options that are directly passed into LaTeX documents.
+- When you have open two or more root files with the same filename, Zettlr will display the containing directory's name as well.
 
 ## Under the Hood